fix duplicate delete

+++ inc/class-s3-uploads.php
@@ -126,11 +126,42 @@ class S3_Uploads {
 					continue;
 				}
 
+				\add_filter(
+					'wp_delete_file',
+					function( $file_wp_wants_to_delete ) use ( $intermediate_file ) {
+						if ( $file_wp_wants_to_delete === $intermediate_file ) {
+							return false;
+						}
+
+						return $file_wp_wants_to_delete;
+					},
+					10,
+					1
+				);
+
 				$deleted[ $intermediate_file ] = true;
 
 				unlink( $intermediate_file );
 			}
 		}
+
+		if ( $deleted[ $file ] ?? false ) {
+			return;
+		}
+
+		\add_filter(
+			'wp_delete_file',
+			function( $file_wp_wants_to_delete ) use ( $file ) {
+				if ( $file_wp_wants_to_delete === $file ) {
+					return false;
+				}
+
+				return $file_wp_wants_to_delete;
+			},
+			10,
+			1
+		);
+
 		unlink( $file );
 	}
